refactor contract flow
