+++ src/app/gigs/contract/index.ts
@@ -1,8 +1,8 @@
-export * from './contract-container';
-export * from './contract-fixed-price';
+export * from './contract-form-container';
+export * from './contract-form-fixed-price';
 export * from './contract-overview';
-export * from './contract-milestones';
-export * from './contract-client-milestone';
+export * from './contract-form-milestones';
+export * from './contract-milestone';
 export * from './contract-not-found';
 export * from './contract-pending-offer';
 export * from './contract-pending-freelancer';
