+++ src/app/gigs/gigs-board.tsx
@@ -1,11 +1,11 @@
 import { GigsColumn } from './gigs-column';
 import { GigsCard } from '@/app/gigs/gigs-card';
-import { GigsComms } from '@/app/gigs/gigs-comms';
+import { GigsDialog } from '@/app/gigs/gigs-dialog';
 import { useAuth, useUpdateOfferStatus } from '@/hooks';
 import { useFetchOffers } from '@/hooks/gigs/useFetchOffers';
 import { GigsContentType, OfferType } from '@/types';
 import { Loader } from 'lucide-react';
-import React, { useState } from 'react';
+import React from 'react';
 
 export interface GigsBoardProps {
   gigsContentType: GigsContentType;
@@ -17,13 +17,6 @@ export const GigsBoard = ({ gigsContentType }: GigsBoardProps) => {
   const [currentOffer, setCurrentOffer] = React.useState<OfferType | undefined>(
     undefined
   );
-  const [freelancerUsername, setFreelancerUsername] = useState<
-    string | undefined
-  >(undefined);
-  const [clientUsername, setClientUsername] = useState<string | undefined>(
-    undefined
-  );
-
   const { isLoading: isFetchOffersLoading, data: offers } = useFetchOffers(
     loggedInUser,
     gigsContentType,
@@ -65,7 +58,7 @@ export const GigsBoard = ({ gigsContentType }: GigsBoardProps) => {
   return (
     <div className='max-w-7xl mx-auto px-4 sm:px-6 lg:px-8'>
       <div className='grid grid-cols-1 md:grid-cols-3 gap-4'>
-        <GigsComms
+        <GigsDialog
           isOpen={isCommsOpen}
           onOpenChange={onGigsCommsOpenChange}
           isClient={gigsContentType === GigsContentType.Sent}
