+++ src/app/gigs/gigs-container.tsx
@@ -15,7 +15,7 @@ export interface GigsContainerProps {
   isClient?: boolean;
 }
 
-export const GigsContainer = ({ isClient = false }) => {
+export const GigsContainer = () => {
   const { user: loggedInUser } = useAuth();
   const { toggleOpenSidebar } = useAppSelector((state) => state.userReducer);
   const [lastChainOp, setLastChainOp] = useLocalStorage<string>(
